Incorporated all the changes(as per peer review)
Change 1 : Added filter for employee types as asked. 
Change 2: Accept the feedback , changed the graph to a histogram
Change 3: Rather than adjusting the scale we have added color to a different kind of response, this created a good visual effect for the kind of responses being captured. The user will instantly know when the repose variable is unique or more than one. 
Change 4: We have removed the second filter from the dashboard and now the whole dashboard will depend on the country selected and the level of the employee (mentioned in graph's subtitle).
Change 5 : Added a default value to the plot
Change 6 : Added imaginary client name